requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.4
python-dotenv==1.0.0

